Add download progress window, refactor everything GB related to another file
