+++ Lang/langs.json
@@ -64,10 +64,10 @@
         "hide_uid_desc": "Hides your User ID text in-game. (Credits: Rokkudayo)",
         "hide_dots_title": "Hide Red Dots",
         "hide_dots_desc": "Hides the notification red dots in-game (Credits: rzwei)",
-        "install_zone_title_zip": "Install Mod from ZIP",
-        "install_zone_choose_zip": "Choose ZIP files",
-        "install_zone_title_folder": "Install Mod from folder",
-        "install_zone_choose_folder": "Choose a folder",
+        "install_zone_title_zip": "Install Mod from Archive",
+        "install_zone_choose_zip": "Choose Archive files",
+        "install_zone_title_folder": "Install Mod from Folder",
+        "install_zone_choose_folder": "Choose a Folder",
         "mod_delete_title": "Deleting Mod",
         "mod_delete_message": "Are you sure you want to delete the following mod?",
         "faq_tooltip": "FAQ Page",
