+++ src/ui/elements.py
@@ -1082,325 +1082,6 @@ class LoadingSpinner(QWidget):
         painter.drawArc(QRectF(-r, -r, r * 2, r * 2), 0, 100 * 16)
         painter.end()
 
-
-class GameBananaMod(QFrame):
-    CARD_W = 160
-    THUMB_H = 90
-
-    _CARD_QSS = """
-        QFrame#GameBananaModCard {
-            background: #13151a;
-            border: 1px solid #2a2d35;
-            border-radius: 10px;
-        }
-        QFrame#GameBananaModCard:hover {
-            border-color: #4493f8;
-            background: #181c24;
-        }
-
-        QLabel#GBStat {
-            color: #8b949e;
-            font-size: 11px;
-            background: transparent;
-            border: none;
-        }
-
-        QLabel#GBName {
-            color: #e6edf3;
-            font-size: 12px;
-            font-weight: 700;
-            background: transparent;
-            border: none;
-        }
-
-        QLabel#GBAuthor {
-            color: #8b949e;
-            font-size: 11px;
-            background: transparent;
-            border: none;
-        }
-
-        QLabel#GBCategory {
-            color: #6e7681;
-            font-size: 10px;
-            background: transparent;
-            border: none;
-        }
-
-        QLabel#GBRatingNSFW {
-            color: #f85149;
-            font-size: 10px;
-            font-weight: 700;
-            background: #3d1c1c;
-            border: 1px solid #6e2a2a;
-            border-radius: 4px;
-            padding: 1px 5px;
-        }
-        QLabel#GBRatingSFW {
-            color: #3fb950;
-            font-size: 10px;
-            font-weight: 700;
-            background: #0f2b18;
-            border: 1px solid #196130;
-            border-radius: 4px;
-            padding: 1px 5px;
-        }
-
-        QPushButton#GBInstallBtn {
-            background: #238636;
-            color: #ffffff;
-            border: none;
-            border-radius: 6px;
-            font-size: 11px;
-            font-weight: 700;
-            padding: 0 8px;
-        }
-        QPushButton#GBInstallBtn:hover  { background: #2ea043; }
-        QPushButton#GBInstallBtn:pressed { background: #1a6b2a; }
-
-        QPushButton#GBOpenBtn {
-            background: #21262d;
-            color: #c9d1d9;
-            border: 1px solid #30363d;
-            border-radius: 6px;
-            font-size: 11px;
-            font-weight: 600;
-            padding: 0 8px;
-        }
-        QPushButton#GBOpenBtn:hover  { background: #30363d; border-color: #4493f8; color: #e6edf3; }
-        QPushButton#GBOpenBtn:pressed { background: #161b22; }
-
-        QFrame#GBSep {
-            background: #21262d;
-        }
-    """
-
-    def __init__(self, mod: NTEMod, parent=None):
-        super().__init__(parent)
-        self.setObjectName("GameBananaModCard")
-        self.setFixedWidth(self.CARD_W)
-        self.setAttribute(Qt.WidgetAttribute.WA_StyledBackground)
-        self.setCursor(Qt.CursorShape.ArrowCursor)
-        self.mod = mod
-        self.setStyleSheet(self._CARD_QSS)
-
-        root = QVBoxLayout(self)
-        root.setContentsMargins(0, 0, 0, 0)
-        root.setSpacing(0)
-        thumbnail_pix = QPixmap()
-        thumbnail_pix.loadFromData(mod.thumbnail)
-
-        thumb = QLabel()
-        thumb.setObjectName("GBThumbnail")
-        thumb.setFixedSize(self.CARD_W, self.THUMB_H)
-        thumb.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        thumb.setCursor(Qt.CursorShape.PointingHandCursor)
-        thumb.setPixmap(_rounded_pixmap(thumbnail_pix, self.CARD_W, self.THUMB_H, radius=10))
-        thumb.setStyleSheet(
-            "border-top-left-radius: 10px;"
-            "border-top-right-radius: 10px;"
-        )
-
-        def _thumb_click(e, pix=thumbnail_pix):
-            if e.button() == Qt.MouseButton.LeftButton and not pix.isNull():
-                show_image(pix, self)
-        thumb.mousePressEvent = _thumb_click
-        root.addWidget(thumb)
-
-        body = QVBoxLayout()
-        body.setContentsMargins(10, 8, 10, 10)
-        body.setSpacing(6)
-        root.addLayout(body)
-
-        name_lbl = QLabel(mod.name)
-        name_lbl.setObjectName("GBName")
-        name_lbl.setWordWrap(True)
-        name_lbl.setMaximumHeight(38)
-        body.addWidget(name_lbl)
-
-        author_row = QHBoxLayout()
-        author_row.setSpacing(4)
-        author_row.setContentsMargins(0, 0, 0, 0)
-
-        author_icon = QLabel()
-        author_icon.setFixedSize(13, 13)
-        _pix_author = QPixmap(resource_path("Bin/Assets/author.png"))
-        if not _pix_author.isNull():
-            author_icon.setPixmap(
-                _pix_author.scaled(13, 13, Qt.AspectRatioMode.KeepAspectRatio,
-                                   Qt.TransformationMode.SmoothTransformation)
-            )
-        author_icon.setStyleSheet("background: transparent; border: none;")
-
-        author_lbl = QLabel(mod.author)
-        author_lbl.setObjectName("GBAuthor")
-
-        author_row.addWidget(author_icon)
-        author_row.addWidget(author_lbl)
-        author_row.addStretch()
-        body.addLayout(author_row)
-
-        sep1 = QFrame()
-        sep1.setObjectName("GBSep")
-        sep1.setFixedHeight(1)
-        body.addWidget(sep1)
-
-        cat_row = QHBoxLayout()
-        cat_row.setSpacing(5)
-        cat_row.setContentsMargins(0, 0, 0, 0)
-
-        cat_text = mod.root_category
-        if mod.sub_category:
-            cat_text = f"{mod.root_category} › {mod.sub_category}" if cat_text else mod.sub_category
-        cat_lbl = QLabel(cat_text or "—")
-        cat_lbl.setObjectName("GBCategory")
-        cat_lbl.setWordWrap(False)
-
-        if mod.is_nsfw:
-            rating_icon_path = resource_path("Bin/Assets/rating_nsfw.png")
-            rating_lbl = QLabel("NSFW")
-            rating_lbl.setObjectName("GBRatingNSFW")
-        else:
-            rating_icon_path = resource_path("Bin/Assets/rating_sfw.png")
-            rating_lbl = QLabel("SFW")
-            rating_lbl.setObjectName("GBRatingSFW")
-
-        rating_icon_lbl = QLabel()
-        rating_icon_lbl.setFixedSize(13, 13)
-        _pix_rating = QPixmap(rating_icon_path)
-        if not _pix_rating.isNull():
-            rating_icon_lbl.setPixmap(
-                _pix_rating.scaled(13, 13, Qt.AspectRatioMode.KeepAspectRatio,
-                                   Qt.TransformationMode.SmoothTransformation)
-            )
-        rating_icon_lbl.setStyleSheet("background: transparent; border: none;")
-
-        cat_row.addWidget(cat_lbl, stretch=1)
-        cat_row.addWidget(rating_icon_lbl)
-        cat_row.addWidget(rating_lbl)
-        body.addLayout(cat_row)
-
-        sep2 = QFrame()
-        sep2.setObjectName("GBSep")
-        sep2.setFixedHeight(1)
-        body.addWidget(sep2)
-
-        stats_row = QHBoxLayout()
-        stats_row.setSpacing(0)
-        stats_row.setContentsMargins(0, 0, 0, 0)
-
-        def _fmt(n: int) -> str:
-            if n >= 1_000_000:
-                return f"{n / 1_000_000:.1f}M"
-            if n >= 1_000:
-                return f"{n / 1_000:.1f}K"
-            return str(n)
-
-        def _stat_widget(icon_path: str, value: int) -> QWidget:
-            w = QWidget()
-            w.setStyleSheet("background: transparent;")
-            h = QHBoxLayout(w)
-            h.setContentsMargins(0, 0, 0, 0)
-            h.setSpacing(3)
-
-            icon_lbl = QLabel()
-            icon_lbl.setFixedSize(12, 12)
-            pix = QPixmap(resource_path(icon_path))
-            if not pix.isNull():
-                icon_lbl.setPixmap(
-                    pix.scaled(12, 12, Qt.AspectRatioMode.KeepAspectRatio,
-                               Qt.TransformationMode.SmoothTransformation)
-                )
-            icon_lbl.setStyleSheet("background: transparent; border: none;")
-
-            val_lbl = QLabel(_fmt(value))
-            val_lbl.setObjectName("GBStat")
-
-            h.addWidget(icon_lbl)
-            h.addWidget(val_lbl)
-            return w
-
-        stats_row.addWidget(_stat_widget("Bin/Assets/views.png",   mod.view_count))
-        stats_row.addStretch()
-        stats_row.addWidget(_stat_widget("Bin/Assets/download.png", mod.download_count))
-        stats_row.addStretch()
-        stats_row.addWidget(_stat_widget("Bin/Assets/like.png",    mod.like_count))
-        body.addLayout(stats_row)
-
-        btn_row = QHBoxLayout()
-        btn_row.setSpacing(6)
-        btn_row.setContentsMargins(0, 0, 0, 0)
-
-        install_btn = QPushButton()
-        install_btn.setObjectName("GBInstallBtn")
-        install_btn.setFixedHeight(28)
-        install_btn.setCursor(Qt.CursorShape.PointingHandCursor)
-        install_btn.setToolTip(t("gamebanana_install_btn") or "Download & Install")
-
-        _dl_icon = QPixmap(resource_path("Bin/Assets/download.png"))
-        if not _dl_icon.isNull():
-            install_btn.setIcon(QIcon(_dl_icon))
-            install_btn.setIconSize(QSize(13, 13))
-        install_btn.setText(t("gamebanana_install_btn") or "Install")
-        install_btn.clicked.connect(self._install)
-
-        open_btn = QPushButton()
-        open_btn.setObjectName("GBOpenBtn")
-        open_btn.setFixedHeight(28)
-        open_btn.setFixedWidth(34)
-        open_btn.setCursor(Qt.CursorShape.PointingHandCursor)
-        open_btn.setToolTip(t("gamebanana_open_btn") or "Open on GameBanana")
-
-        _gb_icon = QPixmap(resource_path("Bin/Assets/marketplace.png"))
-        if not _gb_icon.isNull():
-            open_btn.setIcon(QIcon(_gb_icon))
-            open_btn.setIconSize(QSize(16, 16))
-        open_btn.clicked.connect(self._open_gamebanana)
-
-        btn_row.addWidget(install_btn, stretch=1)
-        btn_row.addWidget(open_btn)
-        body.addLayout(btn_row)
-
-    def _install(self):
-        from src.gamebanana.api import get_mod_files
-        files = get_mod_files(self.mod.id)
-        
-        if len(files) == 1:
-            file = files[0]
-            self.window().__dict__["mod_overlay"].__dict__["gamebanana_install_zone"].install_file(
-                file.name,
-                file.url
-            )
-            return
-        
-        self.overlay = _InstallSelectionOverlay(self.window(), files)
-        self.overlay.show()
-        self.overlay.raise_()
-    
-    def _open_gamebanana(self):
-        url = self.mod.mod_url
-        if not url:
-            url = f"https://gamebanana.com/mods/{self.mod.id}"
-
-        host = self.window()
-        if host is None or host is self:
-            webbrowser.open(url)
-            return
-
-        PopupDialog(
-            parent=host,
-            title="Open GameBanana",
-            message=(
-                f"You are about to open an external link in your browser:\n\n"
-                f"{url}\n\n"
-                "Continue?"
-            ),
-            confirm_text="Open in Browser",
-            cancel_text=t("cancel") or "Cancel",
-            on_confirm=lambda: webbrowser.open(url),
-        )
-
-
 class _ImageViewerOverlay(QWidget):
     def __init__(self, parent: QWidget, pixmap: QPixmap):
         super().__init__(parent)
@@ -1445,139 +1126,3 @@ def show_image(pixmap: QPixmap, parent_widget: QWidget):
     overlay = _ImageViewerOverlay(parent_widget.window(), pixmap)
     overlay.show()
     overlay.raise_()
-
-class _InstallSelectionOverlay(QWidget):
-    def __init__(self, parent: QWidget, files: Optional[List[NTEModFile]]):
-        super().__init__(parent)
-        self.setObjectName("InstallSelectionOverlay")
-        self.setFixedSize(parent.size())
-        self.move(0, 0)
-        self.setAttribute(Qt.WidgetAttribute.WA_StyledBackground)
-        self.setStyleSheet("""
-            QWidget#InstallSelectionOverlay {
-                background: rgba(0, 0, 0, 180);
-            }
-        """)
-        
-        self.card = QFrame(self)
-        self.card.setObjectName("InstallCard")
-        self.card.setFixedWidth(460)
-        self.card.setAttribute(Qt.WidgetAttribute.WA_StyledBackground)
-        self.card.setStyleSheet("""
-            QFrame#InstallCard {
-                background: #161b22;
-                border: 1px solid #30363d;
-                border-radius: 12px;
-            }
-        """)
-        
-        card_layout = QVBoxLayout(self.card)
-        card_layout.setContentsMargins(24, 24, 24, 24)
-        card_layout.setSpacing(16)
-        
-        title = QLabel(t("gamebanana_install_title") or "Select a file to install")
-        title.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        title.setStyleSheet("font-size: 16px; font-weight: bold; color: #e6edf3; background: transparent; border: none;")
-        card_layout.addWidget(title)
-                
-        scroll = QScrollArea()
-        scroll.setWidgetResizable(True)
-        scroll.setStyleSheet("""
-            QScrollArea { border: none; background: transparent; }
-            QScrollBar:vertical { background: #161b22; width: 6px; border-radius: 3px; }
-            QScrollBar::handle:vertical { background: #3d444d; border-radius: 3px; min-height: 20px; }
-        """)
-        
-        files_widget = QWidget()
-        files_widget.setStyleSheet("background: transparent;")
-        files_layout = QVBoxLayout(files_widget)
-        files_layout.setSpacing(8)
-        files_layout.setContentsMargins(0, 0, 0, 0)
-        
-        if files:
-            for file in files:
-                btn = QPushButton(f"{file.name} ({bytes_to_human_readable(file.size)})")
-                btn.setCursor(Qt.CursorShape.PointingHandCursor)
-                btn.setStyleSheet("""
-                    QPushButton {
-                        background: #238636;
-                        color: #ffffff;
-                        border: none;
-                        border-radius: 6px;
-                        font-size: 13px;
-                        font-weight: 600;
-                        padding: 12px 16px;
-                        text-align: left;
-                    }
-                    QPushButton:hover  { background: #2ea043; }
-                    QPushButton:pressed { background: #1a6b2a; }
-                """)
-                btn.clicked.connect(
-                    partial(
-                        parent.__dict__["mod_overlay"].__dict__["gamebanana_install_zone"].install_file,
-                        file.name,
-                        file.url
-                    )
-                )
-                files_layout.addWidget(btn)
-        else:
-            empty_lbl = QLabel(t("gamebanana_install_empty") or "No files available.")
-            empty_lbl.setStyleSheet("color: #8b949e; font-size: 13px;")
-            empty_lbl.setAlignment(Qt.AlignmentFlag.AlignCenter)
-            files_layout.addWidget(empty_lbl)
-            
-        scroll.setWidget(files_widget)
-        
-        files_widget.adjustSize()
-        content_height = files_widget.sizeHint().height()
-        scroll.setFixedHeight(min(content_height, 250))
-        
-        card_layout.addWidget(scroll)
-        
-        btn_row = QHBoxLayout()
-        btn_row.addStretch()
-        cancel_btn = QPushButton(t("cancel") or "Cancel")
-        cancel_btn.setCursor(Qt.CursorShape.PointingHandCursor)
-        cancel_btn.setStyleSheet("""
-            QPushButton {
-                background: #21262d;
-                color: #c9d1d9;
-                border: 1px solid #30363d;
-                border-radius: 6px;
-                font-size: 13px;
-                font-weight: 600;
-                padding: 8px 24px;
-            }
-            QPushButton:hover  { background: #30363d; border-color: #8b949e; color: #e6edf3; }
-            QPushButton:pressed { background: #161b22; }
-        """)
-        cancel_btn.clicked.connect(self._close_overlay)
-        btn_row.addWidget(cancel_btn)
-        
-        card_layout.addLayout(btn_row)
-        
-        self.card.adjustSize()
-        self.card.move(
-            (self.width() - self.card.width()) // 2,
-            (self.height() - self.card.height()) // 2
-        )
-        
-        self.opacity_effect = QGraphicsOpacityEffect(self)
-        self.setGraphicsEffect(self.opacity_effect)
-        self.anim = QPropertyAnimation(self.opacity_effect, b"opacity")
-        self.anim.setDuration(200)
-        self.anim.setStartValue(0)
-        self.anim.setEndValue(1)
-        self.show()
-        self.raise_()
-        self.anim.start()
-
-    def mousePressEvent(self, e):
-        if e.button() == Qt.MouseButton.LeftButton:
-            if not self.card.geometry().contains(e.pos()):
-                self._close_overlay()
-
-    def _close_overlay(self):
-        self.anim.setDirection(QPropertyAnimation.Direction.Backward)
-        self.anim.finished.connect(self.deleteLater)
-        self.anim.start()
