+++ src/ui/mod_manager.py
@@ -2,20 +2,18 @@ import os
 import sys
 import shutil
 import subprocess
-from src.gamebanana.api import NTEMod
-from src.utils import download_file, get_mods_path, resource_path
+from src.gamebanana.window import GameBananaBrowserOverlay, InstallProgressWindow
+from src.utils import get_mods_path, resource_path
 from pathlib import Path
 from PyQt6.QtWidgets import (
-    QWidget, QVBoxLayout, QHBoxLayout, QGridLayout,
-    QPushButton, QLabel, QFrame, QLineEdit,
+    QWidget, QVBoxLayout, QHBoxLayout, QPushButton, QLabel, QFrame, QLineEdit,
     QScrollArea, QFileDialog,
 )
-from PyQt6.QtCore import Qt, QSize, QThread, QObject, pyqtSignal
+from PyQt6.QtCore import Qt, QSize, QObject, pyqtSignal
 from PyQt6.QtGui import QIcon
-from src.styles import GB_STYLE, MOD_MANAGER_STYLE
+from src.styles import MOD_MANAGER_STYLE
 from src.translator import t
-from src.ui.elements import AnimatedToggle, GameBananaMod, ModCard
-from src import config_manager as cfg
+from src.ui.elements import ModCard
 from src.logger import logger
 
 def _ensure_dir(path: Path):
@@ -116,42 +114,10 @@ class _BaseInstallZone(QFrame):
             counter += 1
         return dest
 
-    def _extract_zip(self, src: Path, mods_dir: Path) -> Path:
-        with zipfile.ZipFile(src, "r") as zf:
-            entries = zf.namelist()
-            top_levels = {e.split('/')[0] for e in entries if e.strip('/')}
-            has_single_root = (
-                len(top_levels) == 1 and
-                any(e.startswith(next(iter(top_levels)) + '/') for e in entries)
-            )
-
-            if has_single_root:
-                root_name = next(iter(top_levels))
-                zip_dest = self._unique_dest(mods_dir / root_name)
-                zip_dest.mkdir(parents=True, exist_ok=True)
-                for member in zf.infolist():
-                    member_parts = Path(member.filename).parts[1:]
-                    if not member_parts:
-                        continue
-                    dest_path = zip_dest / Path(*member_parts)
-                    if member.is_dir():
-                        dest_path.mkdir(parents=True, exist_ok=True)
-                    else:
-                        dest_path.parent.mkdir(parents=True, exist_ok=True)
-                        with zf.open(member) as src_f, open(dest_path, 'wb') as dst_f:
-                            dst_f.write(src_f.read())
-            else:
-                zip_dest = self._unique_dest(mods_dir / src.stem)
-                zip_dest.mkdir(parents=True, exist_ok=True)
-                zf.extractall(zip_dest)
-
-        return zip_dest
-
     def _install_paths(self, paths: list[str | Path]):
             paths = [Path(p) for p in paths]
             mods_dir = get_mods_path()
             
-            # Point to your bundled 7za.exe
             seven_zip_path = Path(resource_path("Bin/7z.exe"))
             
             installed_files = []
@@ -212,8 +178,8 @@ class ZipInstallZone(_BaseInstallZone):
         super().__init__(
             mods_dir=mods_dir,
             icon_path="Bin/Assets/install_zip.png",
-            title=t("install_zone_title_zip") or "Install Mod from ZIP",
-            choose_label=t("install_zone_choose_zip") or "Choose ZIP files",
+            title=t("install_zone_title_zip") or "Install Mod from Archive",
+            choose_label=t("install_zone_choose_zip") or "Choose Archive files",
             parent=parent,
         )
 
@@ -221,7 +187,7 @@ class ZipInstallZone(_BaseInstallZone):
         dialog = QFileDialog(self, "Select mod archive files")
         dialog.setFileMode(QFileDialog.FileMode.ExistingFiles)
         dialog.setNameFilters([
-            "Mod archives (*.zip)",
+            "Mod archives (*.zip *.rar *.7z)",
             "All files (*)",
         ])
 
@@ -272,254 +238,12 @@ class GameBananaInstallZone(_BaseInstallZone):
         browser.show()
     
     def install_file(self, filename: str, url: str):
-        file = download_file(filename, url)
-        if file is None:
-            logger.error(f"Failed to download {filename}")
-            return
-        self._install_paths([Path(file)])
+        win = InstallProgressWindow(filename, url)
+        win.install_finished.connect(self.files_installed.emit)
+        win.show()
+        win.start()
+        self._install_win = win
 
-class _ModFetcher(QObject):
-    mod_ready   = pyqtSignal(object)
-    page_done   = pyqtSignal(bool)
-    finished    = pyqtSignal()
-
-    def __init__(self, page: int, force_refresh: bool = False):
-        super().__init__()
-        self._page = page
-        self._force = force_refresh
-        self._cancelled = False
-
-    def cancel(self):
-        self._cancelled = True
-
-    def run(self):
-        from src.gamebanana.api import get_nte_mods
-
-        had_results = False
-
-        def _on_mod(mod):
-            if self._cancelled:
-                return
-            had_results_ref.append(True)
-            self.mod_ready.emit(mod)
-
-        had_results_ref: list = []
-        get_nte_mods(
-            force_refresh=self._force,
-            page=self._page,
-            on_mod_ready=_on_mod,
-        )
-        self.page_done.emit(bool(had_results_ref))
-        self.finished.emit()
-
-
-class GameBananaBrowserOverlay(QFrame):
-    # For handling toggling NSFW mods
-    def handle_toggle(self, checked=None):
-        cfg.set(cfg.Key.SHOW_NSFW_MODS, self.toggle_nsfw_mods.isChecked())
-        
-        if self._fetcher:
-            self._fetcher.cancel()
-
-        self._all_mods.clear()
-        self._current_page = 0
-        self._has_more = True
-        self._loading = False
-
-        while self.gb_grid.count():
-            item = self.gb_grid.takeAt(0)
-            w = item.widget()
-            if w:
-                w.deleteLater()
-
-        self._load_next_page(force_refresh=False)
-            
-    def __init__(self, parent, mod_manager):
-        super().__init__(parent)
-        self.setObjectName("GameBananaBrowserOverlay")
-        self.manager = mod_manager
-
-        self.setGeometry(240, 80, 800, 560)
-        self.setAttribute(Qt.WidgetAttribute.WA_StyledBackground)
-        self.setStyleSheet(GB_STYLE)
-
-        self._all_mods: list = []
-        self._current_page = 0
-        self._has_more = True
-        self._loading = False
-        self._thread: QThread | None = None
-        self._fetcher: _ModFetcher | None = None
-
-        root = QVBoxLayout(self)
-        root.setContentsMargins(0, 0, 0, 0)
-        root.setSpacing(0)
-        header = QFrame()
-        header.setObjectName("GBModManagerHeader")
-        header.setFixedHeight(64)
-        header.setAttribute(Qt.WidgetAttribute.WA_StyledBackground)
-
-        header_layout = QHBoxLayout(header)
-        header_layout.setContentsMargins(28, 0, 20, 0)
-        header_layout.setSpacing(12)
-
-        title_col = QVBoxLayout()
-        lbl_title = QLabel(t("gamebanana_mods") or "GameBanana Mods")
-        lbl_title.setObjectName("GBModManagerTitle")
-        self._lbl_gb_status = QLabel("")
-        self._lbl_gb_status.setObjectName("GBStatus")
-        self._lbl_gb_status.setStyleSheet("color: #8b949e; font-size: 11px;")
-        title_col.addStretch()
-        title_col.addSpacing(24)
-        
-        title_col.addWidget(lbl_title)
-        title_col.addWidget(self._lbl_gb_status)
-        title_col.addStretch()
-        
-        lbl_nsfw_mods = QLabel(t("show_nsfw_mods") or "Show NSFW mods")
-        lbl_nsfw_mods.setStyleSheet("color: #8b949e; font-size: 11px;")
-        
-        self.toggle_nsfw_mods = AnimatedToggle(self)
-        self.toggle_nsfw_mods.setChecked(cfg.get(cfg.Key.SHOW_NSFW_MODS))
-        self.toggle_nsfw_mods.setCursor(Qt.CursorShape.PointingHandCursor)
-        self.toggle_nsfw_mods.setToolTip(t("show_nsfw_mods_tooltip") or "Show NSFW mods")
-
-        btn_clear_cache = QPushButton()
-        btn_clear_cache.setObjectName("SearchActionBtn")
-        btn_clear_cache.setFixedSize(30, 30)
-        btn_clear_cache.setToolTip(t("clear_cache_tooltip") or "Clear mod cache and reload")
-        btn_clear_cache.setCursor(Qt.CursorShape.PointingHandCursor)
-        btn_clear_cache.setIcon(QIcon(resource_path("Bin/Assets/delete.png")))
-        btn_clear_cache.setIconSize(QSize(14, 14))
-        btn_clear_cache.clicked.connect(self._confirm_clear_cache)
-
-        btn_close = QPushButton("✕")
-        btn_close.setObjectName("GBModManagerClose")
-        btn_close.setFixedSize(32, 32)
-        btn_close.setCursor(Qt.CursorShape.PointingHandCursor)
-        btn_close.clicked.connect(self.hide)
-
-        header_layout.addLayout(title_col)
-        header_layout.addStretch()
-        header_layout.addWidget(lbl_nsfw_mods)
-        header_layout.addWidget(self.toggle_nsfw_mods)
-        header_layout.addWidget(btn_clear_cache)
-        header_layout.addWidget(btn_close)
-        root.addWidget(header)
-
-        body = QWidget()
-        body_layout = QVBoxLayout(body)
-        body_layout.setContentsMargins(28, 20, 28, 24)
-        body_layout.setSpacing(16)
-        self.gb_scroll = QScrollArea()
-        self.gb_scroll.setWidgetResizable(True)
-        self.gb_scroll.setFrameShape(QScrollArea.Shape.NoFrame)
-        self.gb_scroll.setStyleSheet("QScrollArea { background: transparent; }")
-        self.gb_container = QWidget()
-        self.gb_container.setObjectName("GBContainer")
-        self.gb_container.setStyleSheet("background: transparent;")
-        self.gb_grid = QGridLayout(self.gb_container)
-        self.gb_grid.setSpacing(10)
-        self.gb_grid.setContentsMargins(0, 0, 0, 0)
-        self.gb_grid.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        self.gb_scroll.setWidget(self.gb_container)
-
-        body_layout.addWidget(self.gb_scroll, 1)
-        root.addWidget(body, 1)
-
-        self.gb_scroll.verticalScrollBar().valueChanged.connect(self._on_scroll)
-
-        self._load_next_page()
-
-    def _load_next_page(self, force_refresh: bool = False):
-        if self._loading or not self._has_more:
-            return
-        self._loading = True
-        self._current_page += 1
-        self._lbl_gb_status.setText(t("loading") or "Loading...")
-        self._thread  = QThread()
-        self._fetcher = _ModFetcher(self._current_page, force_refresh)
-        self._fetcher.moveToThread(self._thread)
-        self._thread.started.connect(self._fetcher.run)
-        self._fetcher.mod_ready.connect(self._on_mod_ready)
-        self._fetcher.page_done.connect(self._on_page_done)
-        self._fetcher.finished.connect(self._thread.quit)
-        self._fetcher.finished.connect(self._fetcher.deleteLater)
-        self._thread.finished.connect(self._thread.deleteLater)
-
-        self._thread.start()
-
-    def _on_mod_ready(self, mod: NTEMod):
-        if mod.is_nsfw and not self.toggle_nsfw_mods.isChecked():
-            return
-        self._all_mods.append(mod)
-        cols = max(1, (self.gb_scroll.width() - 40) // 148)
-        i = len(self._all_mods) - 1
-        card = GameBananaMod(mod)
-        self.gb_grid.addWidget(card, i // cols, i % cols)
-    
-    def _on_page_done(self, had_results: bool):
-        self._lbl_gb_status.setText("")
-        if not had_results:
-            self._has_more = False
-            if not self._all_mods:
-                empty = QLabel(t("no_gamebanana_mods") or "No mods found. Check your connection.")
-                empty.setStyleSheet("color: #8b949e; font-size: 13px; border: none;")
-                empty.setAlignment(Qt.AlignmentFlag.AlignCenter)
-                self.gb_grid.addWidget(empty, 0, 0, 1, 1, Qt.AlignmentFlag.AlignCenter)
-
-        self._loading = False
-        self._check_fill()
-
-    def _check_fill(self):
-        from PyQt6.QtCore import QTimer
-        QTimer.singleShot(0, self._auto_load_if_needed)
-
-    def _auto_load_if_needed(self):
-        if self._loading or not self._has_more:
-            return
-        content  = self.gb_container.sizeHint().height()
-        viewport = self.gb_scroll.viewport().height()
-        if content <= viewport:
-            self._load_next_page()
-
-    def _on_scroll(self, value):
-        if self._loading or not self._has_more:
-            return
-        vbar = self.gb_scroll.verticalScrollBar()
-        if vbar.maximum() > 0 and value >= vbar.maximum() - 50:
-            self._load_next_page()
-
-    def _confirm_clear_cache(self):
-        from src.ui.elements import PopupDialog
-        PopupDialog(
-            parent=self.window(),
-            title=t("clear_cache_title"),
-            message=(
-                t("clear_cache_message")
-            ),
-            confirm_text=t("confirm"),
-            cancel_text=t("cancel"),
-            on_confirm=self._do_clear_cache,
-        )
-
-    def _do_clear_cache(self):
-        from src.gamebanana.api import clear_cache
-        if self._fetcher:
-            self._fetcher.cancel()
-
-        clear_cache()
-        self._all_mods.clear()
-        self._current_page = 0
-        self._has_more = True
-        self._loading = False
-
-        while self.gb_grid.count():
-            item = self.gb_grid.takeAt(0)
-            w = item.widget()
-            if w:
-                w.deleteLater()
-
-        self._load_next_page(force_refresh=True)
 
 
 class ModManagerOverlay(QFrame):
